+++ IPFirewall.m
@@ -24,8 +24,8 @@
 #import "IPFirewall.h"
 
 NSString* const kIPFirewallExecutablePath = @"/sbin/ipfw";
-NSInteger const kIPFirewallRuleSetNumber = 19;
-NSInteger const kIPFirewallRuleStartNumber = 2500;
+int const kIPFirewallRuleSetNumber = 19;
+int const kIPFirewallRuleStartNumber = 2500;
 NSString* const kIPFirewallSelfControlHeader = @"// BEGIN SELFCONTROL BLOCK";
 NSString* const kIPFirewallSelfControlFooter = @"// END SELFCONTROL BLOCK";
